requirements.txt
appnope==0.1.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==21.12b0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
executing==0.8.2
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
flash==1.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
h11==0.13.0
html5lib==1.1
idna==3.3
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
jsonify==0.5
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.6.7
outcome==1.1.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
stack-data==0.1.4
tinysegmenter==0.3
tldextract==3.1.2
tomli==1.2.3
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.1
